refactor: Move employee reminders code to separate module
